+++ src/builder.lisp
@@ -1,6 +1,8 @@
 (uiop:define-package #:docs-builder/builder
   (:use #:cl)
   (:import-from :log4cl)
+  (:import-from #:docs-config
+                #:docs-config)
   (:export #:build))
 (in-package docs-builder/builder)
 
@@ -23,11 +25,28 @@
            rest))
 
 
-(defmethod build :around ((builder t) (system asdf:system) &key &allow-other-keys)
+(defmethod build :around ((builder t) (system asdf:system) &rest rest &key &allow-other-keys)
   (log:info "Building docs for system ~A found at ~A"
             system
             (asdf:system-relative-pathname system ""))
-  (let ((result (call-next-method)))
+  (let* ((primary-system
+           (asdf:find-system (asdf:primary-system-name system)))
+         (special-config (or (docs-config system)
+                             ;; You might be building docs for a subsystem,
+                             ;; but specialize DOCS-CONFIG method on a primary system.
+                             ;; Thats is why here we check both.
+                             (unless (eql system primary-system)
+                               (docs-config primary-system))))
+         (params (append rest
+                         ;; Config defined for ASDF system goes after the method
+                         ;; params because explicit params from REST should have higher priority.
+                         (when special-config
+                           (log:info "Using config ~S" special-config)
+                           special-config)))
+         (result (apply #'call-next-method
+                        builder
+                        system
+                        params)))
     (when (or (null result)
               (not (uiop:directory-exists-p result)))
       (error "BUILD method of (~S ~S) should return a pathname to a directory with HTML docs."